# Performance of bind rows(.id =)