@romainfrancois 2018-12-14
closes #3787