@romainfrancois 2018-12-14
This used to pay for too many write barrier checks and preserve/release of each `Rcpp::String` in the vector.