@romainfrancois 2018-12-14
There is still obviously a difference because it's more work to make the id column than not to, see the `rep(names(data), sizes)` line in the bench results.